feat(agent): move github commands to capability
